adding unit tests
