+++ __tests__/__fixtures__/invalidation/hasCheckIds/2020-10-29T15_31_34.086Z.json
@@ -1,0 +1,21 @@
+{
+  "contextId": {
+    "occurrenceNote": "If it would be clearer in your language, you could say this with an active form. Alternate translation: “that person is excessively proud” (See: [Active or Passive](rc://en/ta/man/translate/figs-activepassive))",
+    "reference": {
+      "bookId": "1ti",
+      "chapter": 6,
+      "verse": 4
+    },
+    "tool": "translationNotes",
+    "groupId": "figs-activepassive",
+    "quote": "τετύφωται",
+    "quoteString": "τετύφωται…νοσῶν",
+    "glQuote": "he is puffed up",
+    "occurrence": 1
+  },
+  "username": "lrsallee",
+  "invalidated": false,
+  "gatewayLanguageCode": "en",
+  "gatewayLanguageQuote": "he is puffed up, … ailing",
+  "modifiedTimestamp": "2020-10-29T15:31:34.086Z"
+}
