+++ __tests__/__fixtures__/invalidation/hasCheckIds/2020-10-30T15_09_34.000Z.json
@@ -1,0 +1,21 @@
+{
+  "contextId": {
+    "occurrenceNote": "Here **he** refers to anyone in general who teaches what is not correct. Alternate translation: “that person is excessively proud” (See: [When Masculine Words Include Women](rc://en/ta/man/translate/figs-gendernotations))",
+    "reference": {
+      "bookId": "1ti",
+      "chapter": 6,
+      "verse": 4
+    },
+    "tool": "translationNotes",
+    "groupId": "figs-gendernotations",
+    "quote": "τετύφωται",
+    "quoteString": "τετύφωται…νοσῶν",
+    "glQuote": "he is puffed up",
+    "occurrence": 1
+  },
+  "username": "lrsallee",
+  "invalidated": false,
+  "gatewayLanguageCode": "en",
+  "gatewayLanguageQuote": "he is puffed up, … ailing",
+  "modifiedTimestamp": "2020-10-30T15:09:34.000Z"
+}
