+++ __tests__/__fixtures__/invalidation/noCheckIds/2020-10-13T15_53_46.554Z.json
@@ -1,0 +1,21 @@
+{
+  "contextId": {
+    "reference": {
+      "bookId": "1ti",
+      "chapter": 1,
+      "verse": 13
+    },
+    "tool": "translationWords",
+    "groupId": "blasphemy",
+    "quote": "βλάσφημον",
+    "strong": [
+      "G09890"
+    ],
+    "occurrence": 1
+  },
+  "username": "lrsallee",
+  "invalidated": false,
+  "gatewayLanguageCode": "en",
+  "gatewayLanguageQuote": "a blasphemer",
+  "modifiedTimestamp": "2020-10-13T15:53:46.554Z"
+}
