+++ __tests__/__fixtures__/invalidation/noCheckIds/2020-10-29T15_37_18.677Z.json
@@ -1,0 +1,30 @@
+{
+  "contextId": {
+    "occurrenceNote": "Paul is referring to his character before he was a Christian. Alternate translation: “a violent person” or “a person who used violence against those who believed in Jesus” (See: [Assumed Knowledge and Implicit Information](rc://en/ta/man/translate/figs-explicit))",
+    "reference": {
+      "bookId": "1ti",
+      "chapter": 1,
+      "verse": 13
+    },
+    "tool": "translationNotes",
+    "groupId": "figs-explicit",
+    "quote": [
+      {
+        "word": "ὄντα",
+        "occurrence": 1
+      },
+      {
+        "word": "βλάσφημον",
+        "occurrence": 1
+      }
+    ],
+    "quoteString": "ὑβριστήν",
+    "glQuote": "an aggressor",
+    "occurrence": 1
+  },
+  "username": "lrsallee",
+  "invalidated": false,
+  "gatewayLanguageCode": "en",
+  "gatewayLanguageQuote": "an aggresor",
+  "modifiedTimestamp": "2020-10-29T15:37:18.677Z"
+}
